as.power4peaks

Convert a DiffBind analyses object (class DBA) into an object of class power4peaks.

as.power4peaks(dba.object, contrast = 1, p.adjust.method = "BH")

Arguments

dba.object

An object of class DBA obtained from DiffBind.

contrast

Numeric value indicating the index corresponding to a contrast stored in the dba.object. Default: 1.

p.adjust.method

String indicating the method to use to correct the p-values. One among: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: BH.

Value

An object of class power4peaks.stats.