compute.power

Compute the power and effect size for the object obtained from get.stats.

compute.power(
  power4peaks.stats,
  sample.size.range = c(0, 20),
  power.threshold = 0.8,
  estimation.method = "congrad",
  distribution.type = NULL,
  df = NULL
)

Arguments

power4peaks.stats

An object of class power4peaks.stats.

sample.size.range

Numeric vector of length 2, indicating minimum and maximum of the range of number of samples in which test the power. Default: c(0,50).

power.threshold

number between 0 and 1 indicating the threshold to plot for the power. Default: 0.8 (80 percent power).

estimation.method

String indicating the method to use to estimate the power (passed to SSPA::sampleSize function). One among: 'deconv', 'congrad', 'tikhonov' or 'ferreira'. Default: "congrad".

distribution.type

String indicating the type of distribution to use; one among: 'norm', 't', 'f', 'chisq'. Default: NULL, automatically defined from the power4peaks.stats object.

df

Numeric value or vector (df1 and df2, required only for F-statistics) indicating the number of degrees of freedom. Default: NULL, automatically defined from the power4peaks.stats object.

Value

An object of class power4peaks.power.